backend/obfuscation/requirements.txt
lief==0.15.1
pycryptodome==3.20.0
